backend-flask/requirements.txt
flask
flask-cors

# Honeycomb
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests 

# AWS X-ray
aws-xray-sdk

# Cloudwatch Logs
watchtower

# Rollbar
blinker
rollbar

# Cognito
Flask-AWSCognito

# Postgres Driver
psycopg[binary]
psycopg[pool]